requirements.txt
crewai
crewai-tools  
langchain-groq
streamlit
python-dotenv
pandas
reportlab
pydantic